fix: identity and wallet creation (#240)
fix: asset preview types
fix: fm search unicode text
fix: feed identity and stamp usage
