+++ src/utils/identity.ts
@@ -1,4 +1,4 @@
-import { BatchId, Bee, NULL_TOPIC, PrivateKey, Reference } from '@ethersphere/bee-js'
+import { BatchId, Bee, Bytes, NULL_TOPIC, PrivateKey, Reference } from '@ethersphere/bee-js'
 import { randomBytes, Wallet } from 'ethers'
 
 import { Identity, IdentityType } from '../providers/Feeds'
@@ -7,7 +7,7 @@ import { LocalStorageKeys } from './localStorage'
 import { uuidV4, waitUntilStampUsable } from '.'
 
 export function generateWallet(): Wallet {
-  const privateKey = randomBytes(PrivateKey.LENGTH).toString()
+  const privateKey = new Bytes(randomBytes(PrivateKey.LENGTH)).toString()
 
   return new Wallet(privateKey)
 }
